- Oauth기능 추가 - Oauth기능으로 본인 글만 삭제 가능하게 기능 추가

+++ pages/api/post/delete.tsx
@@ -1,6 +1,9 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import { connectDB } from "@/util/database";
 import { ObjectId } from "bson";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/pages/api/auth/[...nextauth]";
+import { Db } from "mongodb";
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
 	let id = req.body;
 	const objectId = new ObjectId(id);
@@ -9,6 +12,18 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 		try {
 			const client = await connectDB;
 			const db = client.db("db-stelive");
+
+			const session = await getServerSession(req, res, authOptions);
+
+			// 삭제할 데이터 검색
+			const findData = await getOneData(db);
+
+			// 세션이 있고, 본인이 작성한 글만 삭제 가능
+			if( !session
+				|| session?.user?.email !== findData?.author
+			) {
+				return res.status(403).json('작성자만 삭제할 수 있습니다.');
+			}
 
 			let result = await db.collection('post')
 				.deleteOne({ _id: objectId })
@@ -24,6 +39,21 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 		}
 	}
 
+	// 삭제할 데이터 검색
+	async function getOneData(db: Db) {
+		let result = await db.collection('post')
+			.findOne({ _id: objectId })
+			.then((r) => {
+				console.log(r);
+				return r;
+			})
+			.catch((err) => {
+				return res.status(500).json('MongoDB 처리 오류');
+			});
 
+		return result;
+	}
 }
+
+
